Move all validation code to separate files
