+++ dynestyx/inference/__init__.py
@@ -1,4 +1,5 @@
 __all__ = [
+    "checkers",
     "filters",
     "hmm_filters",
     "integrations",
